Added primitive generation of rooms.
Added primitive generation of rooms. Generates the starting room and the adjacent rooms depending on the adjacent rooms. Fixed a bug where exitsFromRoom was set incorrectly. Fixed a bug where the room rotation was determined regardless of the adjacent rooms. Made exitsFromRoom local to the spawn function of adjacent rooms and the starting room. Fixed a bug where exitsFromRoom was incorrectly changed when the room was turned to the east.