Added the use of pool rooms in room generation.
Added generation of pools with rooms for specific zones (instead of regular rooms, it is planned to generate rooms for a specific zone. 10 zones are planned in the bunker).
Added the ability to select the starting zone.
Added settings for the room: Which zone does the room belong to, whether the prefab room can be used indefinitely, whether the prefab room has been used, whether the room can be used as a starting room.
Now all rooms are preloaded into the zone pool. They have a disabled status.
Now, instead of generating adjacent rooms, the adjacent room is taken from the zone pool and moved to the desired location and its status changes to activated.