The player creates a map by moving.
A bug has been solved where the number of prefab rooms was not enough to create a map from rooms, which caused an exit beyond the array boundary.
Now the room in which the player is located and the rooms that are around this room have the isUsed status. It is made so that when creating an adjacent room, the room in which the player is located does not move and the player does not fall through.
Added a function to check whether the room has already been created. It is used if the player returns or stumbles upon an area with created rooms while passing.
Added a function to turn off distant rooms. If the player goes into an adjacent room, then all old adjacent rooms are turned off (except for the one in which he is located and the one in which he was).
Added a PlayerController that currently only stores the position of the player's previous room.
Added an OnTriggerEnter function to RoomManager. Now if an object with the Player tag enters a room, then the old adjacent rooms disappear, the player's position changes, and new adjacent rooms appear.