Changes in map generation.
Since when moving from one room to the next, two identical adjacent rooms can spawn next to each other, upon entering which, the room in which the player is located disappears and it is transferred to the next room position, it was decided to make 2 types of the game - endless and normal.
The difference between these modes will be that in the endless mode there will be a check of the neighboring 13 rooms when generating a world from rooms, and in the normal mode, all completed rooms will be marked as used and there will be only 4 checks when generating rooms.
Added the creation of a room - "DeadEndRoom". If the pool of rooms runs out, or a situation occurs in which none of the rooms from the room-pool is suitable, then a new room DeadEndRoom will be spawned. It will contain 4 dead-end rooms that will not be connected.
If the player enters the "DeadEndRoom", then the neighboring rooms from this room are not generated. Fixed a bug where there was no reference in "DeadEndRoom" to SpawnRoomManager.